Removed averaging before CSV export
Before the original processing script provided by the authors exported
extracted data to CSV, it averaged every 10 rows of data.
This caused a bug because of the string protocol name feature, which
can't be averaged.

I temporarily removed the averaging. The other option (which the
original authors took) would be to remove the Protocol_name feature.
I'm unsure why the authors are averaging, so I'll reach out to ask them
before I re-implement the averaging.
WARNING: code is in an unstable state since the preprocessing script
doesn't produce the same output as for the dataset created by the
authors.